+++ packages/psychart/example/Plotting Data/src/index.js
@@ -7,9 +7,9 @@ const ps = new Psychart({
     regions: ['Summer (light work)', 'Summer (sitting)', 'Summer (walking)'],
 });
 
-// Append Psychart onto the document and plot data.
+// Append Psychart and the legend onto the document and plot data.
 window.addEventListener('load', () => {
-    document.body.append(ps.getElement());
+    document.body.append(ps.getElement(), ps.getLegend());
     plotData();
 });
 
@@ -52,7 +52,7 @@ function plotData() {
             {
                 advanced: true,
                 line: true,
-                legend: 'Room R1', // Need to assign a legend to connect data points
+                seriesName: 'Room R1', // Need to assign a series name to connect data points
                 time: { start: startTime, end: endTime, now: timeStamp },
             }
         );
@@ -70,7 +70,7 @@ function plotData() {
             {
                 advanced: true,
                 line: true,
-                legend: 'Room R2', // Assign another legend to differentiate data series
+                seriesName: 'Room R2', // Assign another series name to differentiate data series
                 time: { start: startTime, end: endTime, now: timeStamp },
                 gradient: 'Sunset', // Set a non-default gradient type to make it easier to visually differentiate between the two data series
             }