+++ packages/viridis/src/Color.ts
@@ -46,8 +46,20 @@ export class Color {
      * const css = red.toString(); // rgba(255,0,0,100%)
      * ```
      */
-    public toString(): string {
-        return 'rgba(' + this.red + ',' + this.green + ',' + this.blue + ',' + this.alpha + '%)';
+    public toString(type: 'rgb' | 'rgba' | 'hex' | 'hex' | 'hex-transparency' = 'rgba'): string {
+        switch (type) {
+            case 'rgb':
+                return 'rgb(' + this.red + ',' + this.green + ',' + this.blue + ')';
+            case 'rgba':
+                return 'rgba(' + this.red + ',' + this.green + ',' + this.blue + ',' + this.alpha + '%)';
+            case 'hex':
+                return '#' + SMath.toHex(this.red, 2) + SMath.toHex(this.green, 2) + SMath.toHex(this.blue, 2);
+            case 'hex-transparency':
+                const alpha255: number = SMath.clamp(SMath.round2(SMath.translate(this.alpha, 0, 100, 0, 255), 1), 0, 255);
+                return this.toString('hex') + SMath.toHex(alpha255, 2);
+            default:
+                throw new Error('Invalid color type: ' + type);
+        }
     }
     /**
      * Create a new color given a hexadecimal string.
@@ -67,6 +79,6 @@ export class Color {
         if (regex === null) {
             throw new Error('Invalid hexadecimal string: ' + hex);
         }
-        return new Color(parseInt(regex[1], 16), parseInt(regex[2], 16), parseInt(regex[3], 16), parseInt(regex[4] ?? 'FF', 16));
+        return new Color(parseInt(regex[1], 16), parseInt(regex[2], 16), parseInt(regex[3], 16), SMath.translate(parseInt(regex[4] ?? 'FF', 16), 0, 255, 0, 100));
     }
 }