+++ packages/viridis/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.8
+
+- Fix bug where using a hexadecimal color string would inaccurately parse the alpha channel, causing it to (almost) always be fully opaque
+- Create `Gradient.toSVG()` function which creates a `<linearGradient>` element to be appended onto `<defs>`, since SVG elements cannot use the CSS linear gradient styling methods
+- Allow different string representations of a color using a parameter in the `toString()` function:
+    - RGB (red, green, blue)
+    - RGBA (RGB with transparency)
+    - Hexadecimal color code
+    - Hexadecimal with transparency
+- Update unit tests
+
 ## 1.1.7
 
 - This version is functionally identical to the previous, but is published as a test of the updated workflow, with a cleaned up `package.json` file